pyproject.toml
[tool.poetry]
name = "cluo"
version = "0.7.7"
description = "Cluo is a stage-based, composable data stream processing library in Python."
authors = ["Techstars"]
readme = "README.md"
repository = "https://github.com/techstars/cluo"
keywords = []
packages = [
    { include = "cluo" }
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
rich = "^10.15.1"
confluent-kafka = "^1.7.0"
SQLAlchemy = "^1.4.27"
psycopg2-binary = "^2.9.3"
redis = "^4.1.0"
click = "8.0.4"
types-redis = "^4.1.10"
networkx = "^2.6.3"
Unidecode = "^1.3.2"
requests = "^2.27.1"
phonenumbers = "^8.12.41"
types-requests = "^2.28.11"
simple-salesforce = "^1.11.4"
avro = "1.11.3"
lxml = "4.9.3"
backoff = "^2.2.1"
fastavro = "^1.6.1"
dateparser = "^1.1.2"
types-dateparser = "^1.1.4"
PyYAML = "6.0.1"
prometheus-client = "^0.19.0"

[tool.poetry.dev-dependencies]
mypy = "^0.991"
pytest = "^7.2.0"
pytest-sugar = "^0.9.4"
pytest-mock = "^3.6.1"
pytest-docker = "^2.0.1"
pytest-cov = "^3.0.0"
pdoc = "^12.2.0"
interrogate = "^1.5.0"
types-redis = "^4.1.10"
xenon = "0.9.1"
freezegun = "^1.1.0"
types-freezegun = "^1.1.6"
types-requests = "^2.28.11"
types-dateparser = "^1.1.4"
ptvsd = "^4.3.2"
pytest-order = "^1.0.1"
simple-mockforce = "^0.8.0"
coverage = {extras = ["toml"], version = "^5.5.0"}
ruff = "0.4.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
mypy_path = "./cluo:./tests:./examples"
ignore_missing_imports = true
warn_unused_ignores = true
allow_empty_bodies = true
disallow_untyped_calls = true
allow_redefinition = true
warn_no_return = false

[tool.poetry.scripts]
cluo = "cluo.cli.application:cli"

[tool.ruff]
target-version = "py310"
include = ["pyproject.toml", "cluo/**/*.py", "tests/**/*.py"]

[tool.ruff.lint]
select = [
    "F",  # flake8
    "S",  # bandit
    "E",  # pycodestyle / pylint: error
    "W",  # pycodestyle / pylint: warning
    "I"   # isort
]
ignore = [
    "S608",  # hardcoded-sql-expression
    "S101",  # use of assert: we use it in tests of course
    "E501",   # line-too-long: we use ruff format to fix these but there are still a few that make it through
]

[tool.coverage.run]
source = ["cluo"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.html]
directory = "htmlcov"

[tool.pytest.ini_options]
minversion = "6.2"
addopts = "--cov=cluo/"
testpaths = [
    "tests"
]

[tool.interrogate]
ignore-init-method = false
ignore-init-module = true
ignore-magic = true
ignore-semiprivate = true
ignore-private = true
ignore-property-decorators = true
ignore-module = true
ignore-nested-functions = true
ignore-nested-classes = true
ignore-setters = true
fail-under = 100
exclude = ["setup.py", "docs", "build", "tests", "examples", "tools"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*", ".*process_batch", ".*process_record", "_*"]
# possible values: 0 (minimal output), 1 (-v), 2 (-vv)
verbose = 1
quiet = false
whitelist-regex = [""]
color = true
omit-covered-files = false
